[supply] prevent disabling of tracks on promotion (#13351)
* Add option to prevent disabling of tracks on promotion

This PR fixes an odd bug with Supply’s `promote` functionality. If a user is using Supply to promote APKs between tracks, the current code will deactivate the binary in the lower track. This is not always the desired behavior - sometimes, you want to leave the binary in the lower track untouched. This PR keeps the default behavior, but adds a flag to allow users to opt out of this.

This is similar to the `check_superseded_tracks` option that exists for uploading binaries. It turns out this is not checked when we promote, however. In order to avoid any disruption to existing users, it felt easier to add a new flag.

This will fix https://github.com/fastlane/fastlane/issues/13313 if accepted.

* Remove the option and just “do the right thing”

* Revert "Remove the option and just “do the right thing”"

This reverts commit 122346d765900748985ceeedab8b7449d2832f1c.