[docs] Improve instructions for using Charles to debug spaceship (#13890)
<!-- Thanks for contributing to _fastlane_! Before you submit your pull request, please make sure to check the following boxes by putting an x in the [ ] (don't: [x ], [ x], do: [x]) -->

### Checklist
- [x] I've run `bundle exec rspec` from the root directory to see all new and existing tests pass
- [x] I've followed the _fastlane_ code style and run `bundle exec rubocop -a` to ensure the code style is valid
- [x] I've read the [Contribution Guidelines](https://github.com/fastlane/fastlane/blob/master/CONTRIBUTING.md)
- [x] I've updated the documentation if necessary.

### Motivation and Context
ToolsAndDebugging.md and the spaceship docs had duplicated information and made it harder than necessary for someone to link to one location to instruct people on how to use Charles to create a traffic session.

### Description
Move the technical stuff to spaceship, crosslink back from spaceship to the existing Charles instructions. Fix tiny things along the way.