[deliver, pilot] Look up provider IDs programmatically (#13174)
🔑

### Checklist
- [x] I've run `bundle exec rspec` from the root directory to see all new and existing tests pass
- [x] I've followed the _fastlane_ code style and run `bundle exec rubocop -a` to ensure the code style is valid
- [x] I've read the [Contribution Guidelines](https://github.com/fastlane/fastlane/blob/master/CONTRIBUTING.md)
- [x] I've updated the documentation if necessary.

### Motivation and Context
This PR is motivated by issue #10796. When you're on multiple App Store Connect teams, it would be helpful if Pilot and Deliver could detect your iTunes provider short name.

Andrew mentioned four goals in the conversation on #10796:

1. constructing and running this long command from a fastlane action
2. using my existing credentials in the keychain
3. parsing the output into something that can be easily consumed programmatically
4. automatically mapping to developer portal team IDs & iTunes Connect team IDs based on the organization name

This PR implements goals 2, 3, and 4 on that list. I'm not sure about adding the idea of adding a core fastlane action, at least not as part of this PR. However, with the changes to ItunesTransporter described below, it will become possible to make a plugin or custom action that prints out the list of provider names.

I tested this PR by using Pilot to upload an IPA with and without specifying a provider short name.

### Description
This PR modifies three components: Pilot, Deliver, and ItunesTransporter. My goal was to keep Pilot and Deliver completely unchanged for single-team users, but add the power to look up the provider short name programmatically for users who are on more than one App Store Connect team.

#### ItunesTransporter
A new method, `provider_ids`, is added to ItunesTransporter in Fastlane Core. It runs iTMS Transporter with the `-m provider` option and parses the table of provider names returned by Apple. It returns a map of provider short names keyed by team name.

#### Pilot and Deliver
When you're on fewer than two App Store Connect teams, no change.

If you pass an explicit `itc_provider`, no change.

Otherwise, Pilot and Deliver automatically invoke iTMS Transporter with `-m provider`. If one of the provider long names matches the name of the active team in your Tunes session, the corresponding provider short name will be used to upload or download binaries.

Thank you so much for your time!

Co-authored-by: Jan Piotrowski <piotrowski+github@gmail.com>